Update events/molecule.h, events/molecule.cpp
It is much easier and correct to filter a
spurious molecule event when it is detected than
to remove it from the engine later, when it is already
in the events heap.

Molecule events now compute a finite-differences
time derivative when a zero for the inter-center
distance is found. If the derivative is positive, the
collision is rejected and the scan for new collisions
carries on until the end of the bins.

Only the sign of the finite-difference derivative is
used. However, it is important to remark that the time
increment size poses a limit to the maximum speed
particles can spin in our framework.