Update engine, window
Implements basic tag systems. Engine offers a tag and
an untag method that can be used to add and remove
tags from molecules with a given id.

Also modifies window to display only molecules with
a given tag. The example makes a particle appear and
disappear from the window by tagging and untagging it
at regular time intervals.