Major corrections for report::molecule implementation
The previous commit was broken and needed a serous revision. Now
it should be fully functional.
With report::molecule it is possible to assemble any desired lambda for
analyzing and checking the characteristics of a molecule event
whitout altering it.
In main.cpp there is an exemple of its possibilities.