Update dispatcher, engine
It is now possible to subscribe to bumper events
while filtering by specific tag, and to unsubscribe
from bumper events.

This commit includes a hotfix on the detection of
bumper events: like in molecule events, we compute a
finite-differences time derivative of the positions
of colliding atoms, and filter only for approaching
atoms.