Add resetter
Adds resetter class, which is just a helper class for
engine to offer an easy reset interface as a public
member.

Extends molecule with a scale_energy method, which
scales the energy of the molecule to a given target
while keeping the distribution among degrees of
freedom the same.

Resetter class offers a method to scale energy
(without creating issues with the event flow) for
either a single molecule, all the molecules under a
tag, or all the molecules.