Add temperature system for bumpers
Since we want to be able to make thermical walls composed by bumpers of given temperature, this is a basic upgrade of bumper class in order to be able to do so.

This is only for finding the right spot for placing the feature, the physics of the thing right now is completely inexistent.

Moreover, I will probably study a better architectural solution for this thing in the future.